memer-bot/requirements.txt
python-telegram-bot[job-queue]==20.*
asyncpg
requests